finalVersion
Demo of our project, which we developed as two people, using Binary Search Tree and LinkedList algorithms, the algorithms of which were made by me and the interface by my friend. It is a simple application that lists all the works in the museum and draws the route of how to reach the room whose code is entered, as a solution to the problem of rooms located inside each other in some museums and how to reach which work becomes a pain.